+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## 4.2.3
+
+added repro3 orbit option (GFZ multi-GNSS from the earlier days)
+
+Fixed rinex2snr that was crapping out when the last datablock was illegal.
+Files had been made by teqc ----
+
+
 
 ## 4.2.2